Updated scalers to continue incrementing when the 16bits rolls over. The scalers are now unsigned long values. Also added an example for adcGain parameter for completeness

+++ dataItems.cpp
@@ -116,7 +116,7 @@ int OutputEvent::SetPulserNumber(int pulserNumberIn) {
     return pulserNumber;
 }
 
-int OutputEvent::AddScalerEvent(int channel, uint16_t value) {
+int OutputEvent::AddScalerEvent(int channel, unsigned long value) {
     scalerChannels[channel] = value;
     return 0;
 }
