+++ dataItems.h
@@ -31,7 +31,7 @@ public:
     unsigned long eventNumber;
     double adcChannels[128];
     double tdcChannels[128];
-    uint16_t scalerChannels[16];
+    unsigned long scalerChannels[16];
     int adcMultiplicity;
     int tdcMultiplicity;
     unsigned long pulserNumber;
@@ -42,6 +42,6 @@ public:
     int SetADCMultiplicity();
     int SetTDCMultiplicity();
     int SetPulserNumber(int pulserNumberIn);
-    int AddScalerEvent(int channel, uint16_t value);
+    int AddScalerEvent(int channel, unsigned long value);
 
 };
